K.fun {elliptic}	R Documentation

quarter period K

Description

Calculates the K.fun in terms of either m (K.fun()) or k (K.fun.k()).

Usage

K.fun(m, strict=TRUE, maxiter=7)

Arguments

m
Real or complex parameter

strict
Boolean, with default TRUE meaning to return an error if the sequence has not converged exactly, and FALSE meaning to return the partial sum, and a warning

maxiter
Maximum number of iterations

Author(s)

Robin K. S. Hankin

References

R. Coquereaux, A. Grossman, and B. E. Lautrup. Iterative method for calculation of the Weierstrass elliptic function. IMA Journal of Numerical Analysis, vol 10, pp119-128, 1990

Examples

K.fun(0.09)  # AMS-55 give 1.60804862 in example 7 on page 581

# next example not run because: (i), it needs gsl; (ii) it gives a warning.
## Not run: 
K.fun(0.4,strict=F, maxiter=4) - ellint_Kcomp(sqrt(0.4))

## End(Not run)

[Package elliptic version 1.4-0 Index]