P.laurent {elliptic}	R Documentation

Laurent series for elliptic and related functions

Description

Laurent series for various functions

Usage

        P.laurent(z, g=NULL, tol=0, nmax=80)
    Pdash.laurent(z, g=NULL, nmax=80)
    sigma.laurent(z, g=NULL, nmax=8, give.error=FALSE)
sigmadash.laurent(z, g=NULL, nmax=8, give.error=FALSE)
     zeta.laurent(z, g=NULL, nmax=80)

Arguments

z
Primary argument (complex)

g
Vector of length two with g=c(g2,g3)

tol
Tolerance

give.error
In sigma.laurent(), Boolean with default FALSE meaning to return the computed value and TRUE to return the error (as estimated by the sum of the absolute values of the terms along the minor long diagonal of the matrix)

.

nmax
Number of terms used (or, for sigma(), the size of matrix used)

Author(s)

Robin K. S. Hankin

Examples

sigma.laurent(z=1+1i,g=c(0,4))
  

[Package elliptic version 1.4-0 Index]