e16.28.1 {elliptic}	R Documentation

Numerical verification of equations 16.28.1 to 16.28.5

Description

Numerical verification of formulae 16.28.1 to 16.28.5 on p576

Usage

e16.28.1(z, m, ...)
e16.28.2(z, m, ...)
e16.28.3(z, m, ...)
e16.28.4(z, m, ...)
e16.28.5(m, ...)

Arguments

z
Complex number

m
Parameter m

...
Extra arguments passed to theta[1-4]()

Details

Returns the left hand side minus the right hand side of each formula. Each formula documented here is identically zero; nonzero values are returned due to numerical errors and should be small.

Author(s)

Robin K. S. Hankin

References

M. Abramowitz and I. A. Stegun 1965. Handbook of Mathematical Functions. New York, Dover.

Examples

 plot(e16.28.4(z=1:6000,m=0.234))
 plot(abs(e16.28.4(z=1:6000,m=0.234+0.1i)))

[Package elliptic version 1.4-0 Index]