equianharmonic {elliptic}	R Documentation

Special cases of the Weierstrass elliptic function

Description

Gives parameters for the equianharmonic case, the lemniscatic case, and the pseudolemniscatic case.

Usage

equianharmonic(...)
lemniscatic(...)
pseudolemniscatic(...)

Arguments

...
Ignored

Details

These functions return values from section 18.13, p652; 18.14, p658; and 18.15, p662. They use elementary functions (and the gamma function) only, so ought to be more accurate and faster than calling parameters(g=c(1,0)) directly.

Note that the values for the half periods correspond to the general case for complex g2 and g3 so are simple linear combinations of those given in AnS.

One can use parameters("equianharmonic") et seq instead.

Value

Returns a list with the same elements as parameters().

Author(s)

Robin K. S. Hankin

References

M. Abramowitz and I. A. Stegun 1965. Handbook of Mathematical Functions. New York, Dover.

See Also

parameters

Examples

P(z=0.1+0.1212i,params=equianharmonic())

x <- seq(from=-10,to=10,len=200)
z <- outer(x,1i*x,"+")
view(x,x,P(z,params=lemniscatic()),real=FALSE)
view(x,x,P(z,params=pseudolemniscatic()),real=FALSE)
view(x,x,P(z,params=equianharmonic()),real=FALSE)

[Package elliptic version 1.4-0 Index]