latplot {elliptic}	R Documentation

Plots a lattice of periods on the complex plane

Description

Given a pair of basic periods, plots a lattice of periods on the complex plane

Usage

latplot(p, n=10, do.lines=TRUE, ...)

Arguments

p
Vector of length two with first element the first period and second element the second period. Note that p1=2*omega1

n
Size of lattice

do.lines
Boolean with default TRUE meaning to show boundaries between adjacent period parallelograms

...
Extra arguments passed to plot(). See examples section for working use

Author(s)

Robin K. S. Hankin

References

K. Chandrasekharan 1985. Elliptic functions, Springer-Verlag.

Examples

p1 <- c(1,1i)
p2 <- c(2+3i,5+7i)
latplot(p1)
latplot(p2,xlim=c(-4,4),ylim=c(-4,4),n=40)

[Package elliptic version 1.4-0 Index]