lattice {elliptic}	R Documentation

Lattice of complex numbers

Description

Returns a lattice of numbers generated by a given complex basis.

Usage

lattice(p,n)

Arguments

p
Complex vector of length two giving a basis for the lattice

n
size of lattice

Author(s)

Robin K. S. Hankin

Examples

 lattice(c(1+10i,100+1000i),n=2)
plot(lattice(c(1+1i,1.1+1.4i),5))

[Package elliptic version 1.4-0 Index]