massage {elliptic}	R Documentation

Massages numbers near the real line to be real

Description

Returns the Real part of numbers near the real line

Usage

massage(z, tol = 1e-10)

Arguments

z
vector of complex numbers to be massaged

tol
Tolerance

Author(s)

Robin K. S. Hankin

Examples

massage(1+1i)
massage(1+1e-11i)

massage(c(1,1+1e-11i,1+10i))

[Package elliptic version 1.4-0 Index]