mob {elliptic}	R Documentation

Moebius transformations

Description

Moebius transformations

Usage

mob(M, x)
M %mob% x

Arguments

M
2-by-2 matrix of integers

x
vector of values to be transformed

Value

Returns a value with the same attributes as x. Elementwise, if

omitted: see PDF

then mob(M,x) is (ax+b)/(cx+d).

Note

This function does not check for M being having integer elements, nor for the determinant being unity.

Author(s)

Robin K. S. Hankin

References

Wikipedia contributors, "Mobius transformation," Wikipedia, The Free Encyclopedia (accessed February 13, 2011).

See Also

unimodular

Examples

M <- matrix(c(11,6,9,5),2,2)
x <- seq(from=1+1i,to=10-2i,len=6)

M %mob% x
plot(mob(M,x))

[Package elliptic version 1.4-0 Index]