p1.tau {elliptic}	R Documentation

Does the right thing when calling g2.fun() and g3.fun()

Description

Takes vectors and interprets them appropriately for input to g2.fun() and g3.fun(). Not really intended for the end user.

Usage

p1.tau(b)

Arguments

b
Vector of periods

Details

If b is of length two, interpret the elements as omega1 and omega2 respectively.

If a two-column matrix, interpret the columns as omega1 and omega2 respectively.

Otherwise, interpret as a vector of tau=omega1/omega2.

Value

Returns a two-component list:

p1
First period

tau
Period ratio

Author(s)

Robin K. S. Hankin

Examples

 p1.tau(c(1+1i,1.1+23.123i))

[Package elliptic version 1.4-0 Index]