parameters {elliptic}	R Documentation

Parameters for Weierstrass's P function

Description

Calculates the invariants g2 and g3, the e-values e1,e2,e3, and the half periods omega1, omega2, from any one of them.

Usage

parameters(Omega=NULL, g=NULL, description=NULL)

Arguments

Omega
Vector of length two, containing the half periods (omega1,omega2)

g
Vector of length two: (g2,g3)

description
string containing “equianharmonic”, “lemniscatic”, or “pseudolemniscatic”, to specify one of A and S's special cases

Value

Returns a list with the following items:

Omega
A complex vector of length 2 giving the fundamental half periods omega1 and omega2. Notation follows Chandrasekharan: half period omega1 is 0.5 times a (nontrivial) period of minimal modulus, and omega2 is 0.5 times a period of smallest modulus having the property omega2/omega1 not real.

The relevant periods are made unique by the further requirement that Re(omega1)>0, and Im(omega2)>0; but note that this often results in sign changes when considering cases on boundaries (such as real g2 and g3).

Note Different definitions exist for omega3! A and S use omega3=omega2-omega1, while Whittaker and Watson (eg, page 443), and Mathematica, have omega1+omega2+omega3=0

q
The nome. Here, q=exp(pi*i*omega2/omega1).

g
Complex vector of length 2 holding the invariants

e
Complex vector of length 3. Here e1, e2, and e3 are defined by

e1=P(omega1/2), e2=P(omega2/2), e3=P(omega3/2),

where omega3 is defined by ω1+omega2+omega3=0.

Note that the es are also defined as the three roots of x^3-g2*x-g3=0; but this method cannot be used in isolation because the roots may be returned in the wrong order.

Delta
The quantity g2^3-27*g3^2, often denoted Greek capital Delta

Eta
Complex vector of length 3 often denoted by the greek letter eta. Here eta=(eta_1,eta_2,eta_3) are defined in terms of the Weierstrass zeta function with eta_iζ(omega_i) for i=1,2,3.

Note that the name of this element is capitalized to avoid confusion with function eta()

is.AnS
Boolean, with TRUE corresponding to real invariants, as per Abramowitz and Stegun

given
character string indicating which parameter was supplied. Currently, one of “o” (omega), or “g” (invariants)

Author(s)

Robin K. S. Hankin

Examples

 ## Example 6, p665, LHS
 parameters(g=c(10,2+0i))

 ## Example 7, p665, RHS
 a <- parameters(g=c(7,6)) ;  attach(a)
 c(omega2=Omega[1],omega2dash=Omega[1]+Omega[2]*2)

  ## verify 18.3.37:
  Eta[2]*Omega[1]-Eta[1]*Omega[2]   #should be close to pi*1i/2

## from Omega to g and and back;
## following should be equivalentto c(1,1i):
 parameters(g=parameters(Omega=c(1,1i))$g)$Omega 

[Package elliptic version 1.4-0 Index]