theta1dash {elliptic}	R Documentation

Derivatives of theta functions

Description

First, second, and third derivatives of the theta functions

Usage

theta1dash(z, ignore = NULL, m = NULL, q = NULL, give.n = FALSE, maxiter = 30)
theta1dashdash(z, ignore = NULL, m = NULL, q = NULL, give.n = FALSE, maxiter = 30)
theta1dashdashdash(z, ignore = NULL, m = NULL, q = NULL, give.n = FALSE, maxiter = 30)

Arguments

z
Primary complex argument

ignore
Dummy argument to force the user to name the next argument either m or q

m
m as documented in theta1()

q
q as documented in theta1()

give.n
Boolean with default FALSE meaning to return the function evaluation, and TRUE meaning to return a two element list, with first element the function evaluation, and second element the number of iterations used

maxiter
Maximum number of iterations

Details

Uses direct expansion as for theta1() et seq

Author(s)

Robin K. S. Hankin

References

M. Abramowitz and I. A. Stegun 1965. Handbook of Mathematical Functions. New York, Dover

See Also

theta

Examples

m <- 0.3+0.31i
z <- seq(from=1,to=2+1i,len=7)
delta <- 0.001
deriv.numer <- (theta1dashdash(z=z+delta,m=m) - theta1dashdash(z=z,m=m))/delta
deriv.exact <- theta1dashdashdash(z=z+delta/2,m=m)
abs(deriv.numer-deriv.exact)

[Package elliptic version 1.4-0 Index]