quarter period K

Calculates the K.fun in terms of either \(m\) (K.fun()) or \(k\) (K.fun.k()).

Usage

K.fun(m, strict=TRUE, maxiter=7, miniter=3)

Arguments

m

Real or complex parameter

strict

Boolean, with default TRUE meaning to return an error if the sequence has not converged exactly, and FALSE meaning to return the partial sum, and a warning

maxiter

Maximum number of iterations

miniter

Minimum number of iterations to guard against premature exit if an addend is zero exactly

References

R. Coquereaux, A. Grossman, and B. E. Lautrup. Iterative method for calculation of the Weierstrass elliptic function. IMA Journal of Numerical Analysis, vol 10, pp119-128, 1990

Author

Robin K. S. Hankin

Examples

K.fun(0.09)  # AMS-55 give 1.60804862 in example 7 on page 581
#> [1] 1.608049

# next example not run because: (i), it needs gsl; (ii) it gives a warning.
if (FALSE) { # \dontrun{
K.fun(0.4,strict=F, maxiter=4) - ellint_Kcomp(sqrt(0.4))
} # }