Does the Right Thing (tm) when calling g2.fun() and g3.fun()

Takes vectors and interprets them appropriately for input to g2.fun() and g3.fun(). Not really intended for the end user.

Usage

p1.tau(b)

Arguments

b

Vector of periods

Details

If b is of length two, interpret the elements as \(\omega_1\) and \(\omega_2\) respectively.

If a two-column matrix, interpret the columns as \(\omega_1\) and \(\omega_2\) respectively.

Otherwise, interpret as a vector of \(\tau=\omega_1/\omega_2\).

Value

Returns a two-component list:

p1

First period

tau

Period ratio

Author

Robin K. S. Hankin

Examples

 p1.tau(c(1+1i,1.1+23.123i))
#> $p1
#> [1] 1+1i
#> 
#> $tau
#> [1] 12.1115+11.0115i
#>